refactoring of test running
