+++ META-INF/plugin.xml
@@ -46,7 +46,7 @@
       <keyboard-shortcut first-keystroke="alt shift K" keymap="$default"/>
     </action>
 
-    <action id="AtoumRunTests" class="pl.projectspace.idea.plugins.php.atoum.actions.AtoumRunTests"
+    <action id="AtoumRunTests" class="org.atoum.intellij.plugin.atoum.actions.Run"
             icon="/pl/projectspace/idea/plugins/php/atoum/icons/atoum_16_16.png"
             text="run tests">
       <add-to-group group-id="RunMenu" anchor="last"/>
